universal_migration_service/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
clickhouse-connect==0.6.23
PyMySQL==1.1.0
pyodbc==5.0.1
requests==2.31.0
python-dotenv==1.0.0

